add image vae code

+++ serotiny/models/vae/priors/identity_prior.py
@@ -10,6 +10,6 @@ class IdentityPrior(Prior):
 
     def forward(self, z, mode="kl", **kwargs):
         if mode == "kl":
-            return torch.tensor(0)
+            return torch.tensor(0).float()
         else:
             return z
